+++ pints/_mcmc/_summary.py
@@ -6,11 +6,10 @@
 # copyright notice and full license details.
 #
 #
-from __future__ import absolute_import, division
-from __future__ import print_function, unicode_literals
 import numpy as np
 import pints
 import warnings
+
 from tabulate import tabulate