+++ pints/interfaces/stan/_stan.py
@@ -1,12 +1,10 @@
 #
-# Interface for Stan models
+# LogPDF that uses Stan models.
 #
 # This file is part of PINTS (https://github.com/pints-team/pints/) which is
 # released under the BSD 3-clause license. See accompanying LICENSE.md for
 # copyright notice and full license details.
 #
-from __future__ import absolute_import, division
-from __future__ import print_function, unicode_literals
 import numpy as np
 import pystan
 import pints