+++ pints/tests/test_opt_cmaes_bare.py
@@ -17,12 +17,6 @@ from shared import CircularBoundaries
 debug = False
 method = pints.BareCMAES
 
-# Consistent unit testing in Python 2 and 3
-try:
-    unittest.TestCase.assertRaisesRegex
-except AttributeError:
-    unittest.TestCase.assertRaisesRegex = unittest.TestCase.assertRaisesRegexp
-
 
 class TestBareCMAES(unittest.TestCase):
     """